Terminando con CiudadanoController | CitaVacunacion | Vacuna Controller | agregando dependencia validation

+++ src/main/java/com/cris959/Vacunar_te/repository/LaboratorioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LaboratorioRepository extends JpaRepository<Laboratorio, Long> {
+public interface LaboratorioRepository extends JpaRepository<Laboratorio, Integer> {
 
     // Para evitar duplicados legales
     boolean existsByCuit(String cuit);
